fixes elm/core#970
This does not fully solve the problem of detecting case in Unicode, as it can also vary by locale.
This does make the isUpper/Lower and toUpper/Lower functions consistent.

Make Char.isLower and Char.isUpper Unicode-aware
This allows people with non-ASCII alphabets work with `Char.isLower` and `Char.isUpper`. Uses `toUpper` and `toLower`
underneath, which use Javascript's `String.prototype.toLower/UpperCase()`.
The second condition in the functions is there to distinguish between characters that have an upper/lower-case pairing,
and those that don't (`'0' == Char.toLower '0'` but we don't want `isLower '0'` to be true).